add resources
Adds yokubi, which is a lovely short guide. good to smash it out to get an idea of what grammar you will see in the wild, and also as a fast reference for common grammar.

Also adds an anki deck that kinda like bunpro, but free and follows tae kims ordering.